requirements.txt
boto==2.49.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
oauthlib==3.2.0
protobuf==3.15.8
python-dotenv==0.17.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==1.0.1
